+++ src/main/java/forestry/energy/EnergyTransferMode.java
@@ -4,22 +4,10 @@ public enum EnergyTransferMode {
 	EXTRACT, RECEIVE, BOTH, NONE;
 
 	public boolean canExtract() {
-		switch (this) {
-			case EXTRACT:
-			case BOTH:
-				return true;
-			default:
-				return false;
-		}
+		return this == EXTRACT || this == BOTH;
 	}
 
 	public boolean canReceive() {
-		switch (this) {
-			case RECEIVE:
-			case BOTH:
-				return true;
-			default:
-				return false;
-		}
+		return this == RECEIVE || this == BOTH;
 	}
 }