+++ changelog.md
@@ -3,6 +3,7 @@
 - Fix integer alleles always being recessive (#146)
 - Fixed Naturalist chests not closing when KubeJS is installed (#148)
 - Relax the version requirement on Forge to allow launching Forestry with NeoForge, no guarantees of stability (#147)
+- Fix Engine chaining not working
 
 ## Forestry 2.1.1
 - Fix IIndividualHandlerItem methods returning individuals for empty stacks that were decremented